+++ internal/app/go-filecoin/connectors/retrieval_market/fake_api_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"gotest.tools/assert"
 
+	retrievalmarketconnector "github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/connectors/retrieval_market"
 	"github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/paymentchannel"
-	retrievalmarketconnector "github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/retrieval_market_connector"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/types"
 )
