+++ internal/app/go-filecoin/internal/submodule/retrieval_protocol_submodule.go
@@ -9,8 +9,8 @@ import (
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
 	"github.com/libp2p/go-libp2p-core/host"
 
+	retmkt "github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/connectors/retrieval_market"
 	"github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/plumbing/cst"
-	retmkt "github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/retrieval_market_connector"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/piecemanager"
 )
